Add box around Image component
