--- packages/toolpad-app/src/toolpad/AppEditor/PageEditor/RenderPanel/index.ts
@@ -1,3 +1,0 @@
-import RenderPanel from './RenderPanel';
-
-export default RenderPanel;
