+++ packages/toolpad-components/src/Image.tsx
@@ -1,4 +1,4 @@
-import { Box, Skeleton, SxProps } from '@mui/material';
+import { Box, Skeleton, SxProps, BoxProps } from '@mui/material';
 import * as React from 'react';
 import { createComponent } from '@mui/toolpad-core';
 
@@ -10,9 +10,21 @@ export interface ImageProps {
   height: number;
   loading?: boolean;
   fit: 'contain' | 'cover' | 'fill' | 'none' | 'scale-down';
+  alignItems?: BoxProps['alignItems'];
+  justifyContent?: BoxProps['justifyContent'];
 }
 
-function Image({ sx: sxProp, src, width, height, alt, loading: loadingProp, fit }: ImageProps) {
+function Image({
+  sx: sxProp,
+  src,
+  width,
+  height,
+  alt,
+  loading: loadingProp,
+  fit,
+  alignItems,
+  justifyContent,
+}: ImageProps) {
   const sx: SxProps = React.useMemo(
     () => ({
       ...sxProp,
@@ -32,21 +44,29 @@ function Image({ sx: sxProp, src, width, height, alt, loading: loadingProp, fit 
   const loading = loadingProp || imgLoading;
 
   return (
-    <Box sx={sx}>
-      {loading ? <Skeleton variant="rectangular" width={width} height={height} /> : null}
-      <Box
-        component="img"
-        src={src}
-        alt={alt}
-        sx={{
-          width: '100%',
-          height: '100%',
-          objectFit: fit,
-          position: 'absolute',
-          visibility: loading ? 'hidden' : 'visible',
-        }}
-        onLoad={handleLoad}
-      />
+    <Box
+      sx={{
+        display: 'flex',
+        alignItems,
+        justifyContent,
+      }}
+    >
+      <Box sx={sx}>
+        {loading ? <Skeleton variant="rectangular" width={width} height={height} /> : null}
+        <Box
+          component="img"
+          src={src}
+          alt={alt}
+          sx={{
+            width: '100%',
+            height: '100%',
+            objectFit: fit,
+            position: 'absolute',
+            visibility: loading ? 'hidden' : 'visible',
+          }}
+          onLoad={handleLoad}
+        />
+      </Box>
     </Box>
   );
 }
@@ -77,6 +97,24 @@ export default createComponent(Image, {
     loading: {
       typeDef: { type: 'boolean' },
       defaultValue: false,
+    },
+    alignItems: {
+      typeDef: {
+        type: 'string',
+        enum: ['start', 'center', 'end', 'space-between', 'space-around', 'space-evenly'],
+      },
+      label: 'Vertical alignment',
+      control: { type: 'VerticalAlign' },
+      defaultValue: 'center',
+    },
+    justifyContent: {
+      typeDef: {
+        type: 'string',
+        enum: ['start', 'center', 'end', 'space-between', 'space-around', 'space-evenly'],
+      },
+      label: 'Horizontal alignment',
+      control: { type: 'HorizontalAlign' },
+      defaultValue: 'start',
     },
     sx: {
       typeDef: { type: 'object' },
